requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
psycopg2-binary==2.9.9
pandas==2.1.4

# Environment and Configuration
python-dotenv==1.0.0

# AI and Language Models
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.0

# Vector Database
chromadb==0.4.18

# Data Processing
numpy==1.24.4

# HTTP and Async
httpx==0.25.2
aiofiles==23.2.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1